requirements.txt
django
djangorestframework
argon2-cffi
djangorestframework-simplejwt
psycopg2-binary
stripe
django-environ